refactor: improve error handling in deleteClientById method

+++ api_facturas/src/main/java/com/api_facturas/Clientes/service/ClienteService.java
@@ -34,10 +34,16 @@ public class ClienteService {
 
     public void deleteClientById(Integer clienteId) {
         try {
-            clienteRepository.deleteById(clienteId);
+            if (clienteRepository.existsById(clienteId)) {
+                clienteRepository.deleteById(clienteId);
+            } else {
+                throw new RuntimeException("No se encontró el cliente con ID " + clienteId);
+            }
         } catch (Exception e) {
             if (e.getMessage().contains("foreign key constraint fails")) {
                 throw new RuntimeException("No se pudo eliminar el cliente con ID " + clienteId + ". Esto puede deberse a que hay facturas asociadas a este cliente. Por favor, elimine primero las facturas asociadas y luego intente nuevamente.");
+            } else if (e.getMessage().contains("No se encontró el cliente")) {
+                throw new RuntimeException("No se encontró el cliente con ID " + clienteId);
             }
             throw new RuntimeException("No se pudo eliminar el cliente");
         }
